COS-285 - Add folder scan to the capabilities.
This is the first re-factoring of the code to enable not just docker image
analysis, but also discover files in a given folder.

The new sub-command is activated by

```
coguard folder-scan <PATH_TO_FOLDER>
```

If no path is supplied, the current working directory is being used.

This feature will for now be not part of the free offering, although
we may revisit this at some point.

The output format and options remain the same as for the Docker image scan.

On the back-end, a cluster named after the folder will be created, and all discovered
configuration files will be uploaded there.