Hotfix versions (#69)
hotfix: Fix the issues with the mac

There were multiple problems: One being that user ids may be already used. The other one was that M1s have a different platform per default, and the versions of terraform that we need were not included in the rpm package for that platform.