COS-355 - Fixing problems with symlinks
It is important to, when we need to open files to check for integrity,
to check if they are not symlinks. And we should only open it if it is in
the current context.

Furthermore, we are fixing the case, where a Docker call times out
for an included image.